python/requirements.txt
mysql-connector-python
fastapi
uvicorn[standard]